refactor: capitalize all classes
