+++ Terrario.pyw
@@ -1,4 +1,4 @@
-from Classes.game import Game
+from Classes.Game import Game
 
 if __name__ == "__main__":
     Game()
